implement Functions in BL

+++ dotNet5782_4228_1070/BL/BL/BLfunc.cs
@@ -21,7 +21,7 @@ namespace BL
             }
         }
         
-        public static double distance(float x1, float y1, float x2, float y2)
+        public static double distance(double x1, double y1, double x2, double y2)
         {
             double d = Math.Pow((Math.Pow(x2 - x1, 2) + Math.Pow(y2 - y1, 2)), 0.5);
             return d;
